# specs/001-administration/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Echoroo Administration API
  description: Authentication, user management, and project management endpoints
  version: 1.0.0
  contact:
    name: Echoroo Team
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /admin/users:
    get:
      tags: [admin]
      summary: List all users
      description: Get paginated list of all users (superuser only)
      operationId: adminListUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          description: Search by email or display name
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListResponse'
        '401':
          description: Not authenticated
        '403':
          description: Not superuser
  /admin/users/{userId}:
    patch:
      tags: [admin]
      summary: Update user
      description: Update user status (superuser only)
      operationId: adminUpdateUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Cannot disable last superuser
        '401':
          description: Not authenticated
        '403':
          description: Not superuser
        '404':
          description: User not found
  /admin/settings:
    get:
      tags: [admin]
      summary: Get system settings
      description: Get all system settings (superuser only)
      operationId: getSystemSettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/SystemSettingResponse'
        '401':
          description: Not authenticated
        '403':
          description: Not superuser

    patch:
      tags: [admin]
      summary: Update system settings
      description: Update system settings (superuser only)
      operationId: updateSystemSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemSettingsUpdateRequest'
      responses:
        '200':
          description: Settings updated
        '401':
          description: Not authenticated
        '403':
          description: Not superuser

  # ============================================
  # Setup Endpoints
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AdminUserUpdateRequest:
      type: object
      properties:
        is_active:
          type: boolean
        is_superuser:
          type: boolean
        is_verified:
          type: boolean
    SystemSettingsUpdateRequest:
      type: object
      properties:
        registration_mode:
          type: string
          enum: [open, invitation]
        allow_registration:
          type: boolean
        session_timeout_minutes:
          type: integer
          minimum: 5
          maximum: 1440
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        organization:
          type: string
        is_active:
          type: boolean
        is_superuser:
          type: boolean
        is_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
...
    AdminUserListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    SystemSettingResponse:
      type: object
      properties:
        key:
          type: string
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object
        value_type:
          type: string
          enum: [string, number, boolean, json]
        description:
          type: string
        updated_at:
          type: string
          format: date-time
...